python/requirements.txt
PyJWT==2.6.0
python-dotenv==0.21.1
cryptography==41.0.6